refactored Building and Market into consensus file
