+++ src/energyplus/Makefile.am
@@ -26,6 +26,6 @@ LDADD += $(ZMQ_LIBS)
 
 bin_PROGRAMS = eplus_agent eplus_agent_helics test_comm
 eplus_agent_SOURCES = eplus_agent.cpp jsoncpp.cpp
-eplus_agent_helics_SOURCES = eplus_agent_helics.cpp jsoncpp.cpp
-test_comm_SOURCES = test_comm.cpp jsoncpp.cpp
+eplus_agent_helics_SOURCES = eplus_agent_helics.cpp consensus.cpp jsoncpp.cpp
+test_comm_SOURCES = test_comm.cpp consensus.cpp jsoncpp.cpp
